more test coverage for lists
